Add GreatCircleLayer Demo
Add layer docs Catalog

remove debugging code
