+++ docs/layers/great-circle-layer.md
@@ -1,3 +1,4 @@
+<!-- INJECT:"GreatCircleLayerDemo" -->
 
 <p class="badges">
   <img src="https://img.shields.io/badge/64--bit-support-blue.svg?style=flat-square" alt="64-bit" />
